Update boot_command and boot_wait
Boot command is now a tad bit shorter.
Thanks to @nickcharlton for finding this out! The keyword "autoinstall" needs to be present.

Boot_wait needs to be long enough for the VM to have actually started, but short enough to keep the ISO's bootloader from executing the installer.
Here we may need a more robust way of dropping into the grub command line...